# docs/swagger.yaml
basePath: /
definitions:
  internal_handler.HealthCheckResponse:
    properties:
      instance_id:
        example: "12345678"
        type: string
      message:
        example: OK
        type: string
      service_name:
        example: bookmark
        type: string
    type: object
  lesson01-ebvn_internal_handler_dto.ShortenReq:
    properties:
      exp:
        type: integer
      url:
        type: string
    required:
    - url
    type: object
info:
  contact: {}
  description: API document for bookMark API
  title: bookMark API
  version: "1.0"
paths:
  /health-check:
    get:
      consumes:
      - application/json
      description: Get network health check for bookmark service
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/internal_handler.HealthCheckResponse'
      summary: Get serviceName and instance_id
      tags:
      - healthCheck
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      description: Retrieve the original URL from the short code and redirect the
        client to it
      parameters:
      - description: Short URL Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to original URL
          schema:
            type: string
        "400":
          description: code is required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: url not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Redirect to original URL
      tags:
      - Redirect
  /v1/links/shorten:
    post:
      description: Accepts a long URL and an expiration time, then generates a shortened
        key.
      parameters:
      - description: URL to be shortened and expiration time
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lesson01-ebvn_internal_handler_dto.ShortenReq'
      responses:
        "200":
          description: Success
          schema:
            additionalProperties: true
            type: object
      summary: ShorttenURL
      tags:
      - shortenURL
swagger: "2.0"